Cleanup and fix acceptance tests (#16)
* Cleanup and fix acceptance tests

The biggest change here is to remove the need for setting up an
inventory or group before running acceptance tests. In addition, a new
function has been added to the acceptance tests for the job resource to
ensure the associated inventory resource is only deleted after the job
has finished running. Without this, the destroy step will fail. The
supplied job template *must* be set to require an inventory on launch,
otherwise the inventory that's passed is not used.

* Increase timeout and disable test caching