Restructures docs to make better use of templates (#17)
This renames the examples so that they can be automatically included in
the docs by tfplugindocs. Some default templates have been added for
resources and data sources. If there's a need to customize the doc page
for a particular resource, they can be copied and edited, as has been
done for the job resource.