[AAP-54793] Add event stream datasource tests (#160)
Adds tests for event stream datasource Metadata/Schema/Configure
functions.

Adds the URL field to BaseEdaDataSource schema because all supporting
structs and functions expect it to be there.

Adds an acceptance test for retrieving the EDA event stream post
url. This also changes the `testing/playbook.yml` to create an EDA
credential and event stream for running the acceptance tests.

---------

Co-authored-by: Daniel Quackenbush <25692880+danquack@users.noreply.github.com>
Co-authored-by: Dan Leehr <dleehr@users.noreply.github.com>